# openapi.yaml
openapi: 3.1.0
info:
  title: LeanStock Inventory API
  description: LeanStock backend with JWT auth/RBAC, email verification, password
    reset, multi-tenant catalog, atomic inventory transfers, Redis reservations, Gmail
    OAuth2 email delivery, and scheduled dead-stock decay.
  version: 1.0.0
paths:
...
  /v1/auth/register:
    post:
      tags:
      - Authentication
      summary: Register
      operationId: register_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Login
      operationId: login_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh
      operationId: refresh_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/auth/logout:
    post:
      tags:
      - Authentication
      summary: Logout
      operationId: logout_v1_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/auth/verify-email:
    post:
      tags:
      - Authentication
      summary: Verify Email Endpoint
      operationId: verify_email_endpoint_v1_auth_verify_email_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/auth/resend-verification:
    post:
      tags:
      - Authentication
      summary: Resend Verification Endpoint
      operationId: resend_verification_endpoint_v1_auth_resend_verification_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/auth/password-reset/request:
    post:
      tags:
      - Authentication
      summary: Password Reset Request Endpoint
      operationId: password_reset_request_endpoint_v1_auth_password_reset_request_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/auth/password-reset/confirm:
    post:
      tags:
      - Authentication
      summary: Password Reset Confirm Endpoint
      operationId: password_reset_confirm_endpoint_v1_auth_password_reset_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/products:
    get:
      tags:
      - Products
      summary: Read Products
      operationId: read_products_v1_products_get
      security:
      - HTTPBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      - name: is_active
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Products
      summary: Create Product Endpoint
      operationId: create_product_endpoint_v1_products_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/products/{product_id}:
    get:
      tags:
      - Products
      summary: Read Product
      operationId: read_product_v1_products__product_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/inventory:
    get:
      tags:
      - Inventory
      summary: Read Inventory
      operationId: read_inventory_v1_inventory_get
      security:
      - HTTPBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: warehouse_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Warehouse Id
      - name: decay_status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(normal|liquidating)$
          - type: 'null'
          title: Decay Status
      - name: low_stock_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Low Stock Only
      - name: sku
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sku
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryPageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/inventory/adjust:
    post:
      tags:
      - Inventory
      summary: Adjust Inventory Endpoint
      operationId: adjust_inventory_endpoint_v1_inventory_adjust_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryAdjustRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/inventory/reserve:
    post:
      tags:
      - Inventory
      summary: Reserve Inventory Endpoint
      operationId: reserve_inventory_endpoint_v1_inventory_reserve_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InventoryAdjustRequest:
      properties:
        variant_id:
          type: string
          format: uuid
          title: Variant Id
        warehouse_id:
          type: string
          format: uuid
          title: Warehouse Id
        quantity_delta:
          type: integer
          title: Quantity Delta
          examples:
          - -3
        reason:
          type: string
          enum:
          - damage
          - theft
          - surplus
          - count_correction
          - other
          title: Reason
        note:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Note
      type: object
      required:
      - variant_id
      - warehouse_id
      - quantity_delta
      - reason
      title: InventoryAdjustRequest
    InventoryItemResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        warehouse_id:
          type: string
          format: uuid
          title: Warehouse Id
        quantity:
          type: integer
          title: Quantity
        reorder_threshold:
          type: integer
          title: Reorder Threshold
        is_low_stock:
          type: boolean
          title: Is Low Stock
        decay_status:
          type: string
          enum:
          - normal
          - liquidating
          title: Decay Status
        current_price:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Current Price
        discount_pct:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Discount Pct
        last_sold_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Sold At
        variant:
          $ref: '#/components/schemas/InventoryVariantView'
      type: object
      required:
      - id
      - warehouse_id
      - quantity
      - reorder_threshold
      - is_low_stock
      - decay_status
      - current_price
      - discount_pct
      - last_sold_at
      - variant
      title: InventoryItemResponse
    InventoryPageResponse:
      properties:
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        has_more:
          type: boolean
          title: Has More
        total_count:
          type: integer
          title: Total Count
        data:
          items:
            $ref: '#/components/schemas/InventoryItemResponse'
          type: array
          title: Data
      type: object
      required:
      - next_cursor
      - has_more
      - total_count
      - data
      title: InventoryPageResponse
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
          examples:
          - admin@arzanshop.kz
        password:
          type: string
          minLength: 8
          title: Password
          examples:
          - Secur3P@ss!
      type: object
      required:
      - email
      - password
      title: LoginRequest
    LogoutRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
          examples:
          - eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      type: object
      required:
      - refresh_token
      title: LogoutRequest
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
          examples:
          - If the account exists, an email has been queued.
      type: object
      required:
      - message
      title: MessageResponse
    PasswordResetConfirmRequest:
      properties:
        token:
          type: string
          minLength: 32
          title: Token
          examples:
          - password-reset-token-from-link
        new_password:
          type: string
          minLength: 8
          title: New Password
          examples:
          - N3wSecur3P@ss!
      type: object
      required:
      - token
      - new_password
      title: PasswordResetConfirmRequest
    PasswordResetRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
          examples:
          - owner@arzanshop.kz
      type: object
      required:
      - email
      title: PasswordResetRequest
    ProductCreateRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 2
          title: Name
          examples:
          - Classic T-Shirt
        category:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Category
          examples:
          - Clothing
        unit_of_measure:
          type: string
          enum:
          - pcs
          - kg
          - pack
          title: Unit Of Measure
          default: pcs
        variants:
          items:
            $ref: '#/components/schemas/ProductVariantCreateRequest'
          type: array
          minItems: 1
          title: Variants
      type: object
      required:
      - name
      - variants
      title: ProductCreateRequest
    ProductPageResponse:
      properties:
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        has_more:
          type: boolean
          title: Has More
        total_count:
          type: integer
          title: Total Count
        data:
          items:
            $ref: '#/components/schemas/ProductResponse'
          type: array
          title: Data
      type: object
      required:
      - next_cursor
      - has_more
      - total_count
      - data
      title: ProductPageResponse
    ProductResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        name:
          type: string
          title: Name
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        unit_of_measure:
          type: string
          title: Unit Of Measure
        is_active:
          type: boolean
          title: Is Active
        variants:
          items:
            $ref: '#/components/schemas/ProductVariantResponse'
          type: array
          title: Variants
      type: object
      required:
      - id
      - tenant_id
      - name
      - category
      - unit_of_measure
      - is_active
      - variants
      title: ProductResponse
    ProductVariantCreateRequest:
      properties:
        sku:
          type: string
          maxLength: 100
          minLength: 1
          title: Sku
          examples:
          - TSHIRT-RED-L
        color:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Color
        size:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Size
        barcode:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Barcode
        cost_price:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d{0,2}0*$
          title: Cost Price
          examples:
          - 2100.0
        selling_price:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d{0,2}0*$
          title: Selling Price
          examples:
          - 4990.0
        liquidation_floor_price:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d{0,2}0*$
          title: Liquidation Floor Price
          examples:
          - 2990.0
      type: object
      required:
      - sku
      - cost_price
      - selling_price
      - liquidation_floor_price
      title: ProductVariantCreateRequest
...
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
          examples:
          - eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      type: object
      required:
      - refresh_token
      title: RefreshRequest
    RegisterRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
          examples:
          - owner@arzanshop.kz
        password:
          type: string
          minLength: 8
          title: Password
          examples:
          - Secur3P@ss!
        name:
          type: string
          maxLength: 255
          minLength: 2
          title: Name
          examples:
          - Aibek Dzhaksybekov
        role:
          $ref: '#/components/schemas/UserRole'
          default: tenant_admin
          examples:
          - tenant_admin
        tenant_name:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Tenant Name
          examples:
          - Arzan Shop
        tenant_slug:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Tenant Slug
          examples:
          - arzan-shop
      type: object
      required:
      - email
      - password
      - name
      title: RegisterRequest
    ResendVerificationRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
          examples:
          - owner@arzanshop.kz
      type: object
      required:
      - email
      title: ResendVerificationRequest
    ReservationRequest:
      properties:
        variant_id:
          type: string
          format: uuid
          title: Variant Id
        warehouse_id:
          type: string
          format: uuid
          title: Warehouse Id
        quantity:
          type: integer
          minimum: 1.0
          title: Quantity
        order_reference:
          type: string
          maxLength: 100
          minLength: 1
          title: Order Reference
      type: object
      required:
      - variant_id
      - warehouse_id
      - quantity
      - order_reference
      title: ReservationRequest
    ReservationResponse:
      properties:
        reservation_id:
          type: string
          title: Reservation Id
        expires_at:
          type: string
          format: date-time
          title: Expires At
      type: object
      required:
      - reservation_id
      - expires_at
      title: ReservationResponse
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          const: bearer
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
          examples:
          - 1800
      type: object
      required:
      - access_token
      - refresh_token
      - expires_in
      title: TokenResponse
...
    UserResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        email:
          type: string
          format: email
          title: Email
        name:
          type: string
          title: Name
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          title: Is Active
        email_verified_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Email Verified At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - tenant_id
      - email
      - name
      - role
      - is_active
      - email_verified_at
      - created_at
      title: UserResponse
    UserRole:
      type: string
      enum:
      - super_admin
      - tenant_admin
      - warehouse_manager
      - analyst
      title: UserRole
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VerifyEmailRequest:
      properties:
        token:
          type: string
          minLength: 32
          title: Token
          examples:
          - email-verify-token-from-link
      type: object
      required:
      - token
      title: VerifyEmailRequest
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer